+++ rentrix-server/src/main/java/com/rentrix/rentrixserver/service/impl/ReviewServiceImpl.java
@@ -2,41 +2,38 @@ package com.rentrix.rentrixserver.service.impl;
 
 import com.rentrix.rentrixserver.dto.ReviewDto;
 import com.rentrix.rentrixserver.entity.Review;
+import com.rentrix.rentrixserver.exception.ApiException;
 import com.rentrix.rentrixserver.mapper.ReviewMapper;
 import com.rentrix.rentrixserver.repository.ReviewRepository;
 import com.rentrix.rentrixserver.service.ReviewService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-
-@RequiredArgsConstructor
 @Service
+@RequiredArgsConstructor
 public class ReviewServiceImpl implements ReviewService {
-	private final ReviewMapper reviewMapper;
+	
 	private final ReviewRepository reviewRepository;
 	
-	public List<ReviewDto> getAllReviews() {
-		List<Review> reviews = reviewRepository.findAll();
-		List<ReviewDto> reviewDtos = new ArrayList<>();
-		reviews.forEach(review -> reviewDtos.add(reviewMapper.toDto(review)));
-		return reviewDtos;
+	@Override
+	public Page<ReviewDto> getAllReviews(Pageable pageable) {
+		return reviewRepository.findAll(pageable).map(ReviewMapper::toDto);
 	}
 	
+	@Override
 	public ReviewDto getReviewById(Long id) {
-		Review review = reviewRepository.findById(id).orElseThrow();
-		return reviewMapper.toDto(review);
+		Review review = reviewRepository.findById(id)
+												  .orElseThrow(() -> ApiException.notFound("Review not found with id: " + id));
+		return ReviewMapper.toDto(review);
 	}
 	
-	public ReviewDto createReview(ReviewDto reviewDto) {
-		Review review = reviewMapper.toEntity(reviewDto);
-		reviewRepository.save(review);
-		return reviewMapper.toDto(review);
-	}
-	
+	@Override
 	public String deleteReview(Long id) {
-		reviewRepository.findById(id).orElseThrow();
+		if (!reviewRepository.existsById(id)) {
+			throw ApiException.notFound("Review not found with id: " + id);
+		}
 		reviewRepository.deleteById(id);
 		return "Review with id: " + id + " has been deleted";
 	}
